% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{ex}
\alias{ex}
\title{Reading in extdata from package}
\usage{
ex(dta)
}
\arguments{
\item{dta}{file to read (.csv, .xlsx, .shp, etc.)}
}
\value{
path to file
}
\description{
Reading in extdata from package
}
\examples{
read.csv(ex("sierra/sierraStations.csv"))
}
