% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sierraData}
\alias{sierraData}
\title{Sierra monthly climate data}
\format{
A data frame with 82 entries and 7 variables selected and renamed \describe{
\item{NAME}{Station name}
\item{MONTH}{YYYYMM}
\item{ELEVATION}{Elevation in meters}
\item{LATITUDE}{Latitude in decimal degrees}
\item{LONGITUDE}{Longitude in decimal degrees}
\item{PRECIPITATION}{February Average Precipitation in mm}
\item{TEMPERATURE}{Febrary Average Temperature in degrees C}
}
}
\source{
\url{https://www.ncdc.noaa.gov/}
}
\usage{
sierraData
}
\description{
Fewer variables than above
}
\keyword{datasets}
