% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{strsplit1}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{string to split}

\item{split}{character to split with}
}
\value{
split into a vector
}
\description{
strsplit1
}
